backend/requirements.txt
google-genai>=0.8.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
